Time Entry Limits
Improved time entry limits to exclude any entry below 1 and above 12 in the hour field or 0 and 59 in the minute field. When using the arrows, the user cannot go outside of this range and when typing, the value is snapped to the closest value (usually the minimum or the maximum) when clicked outside of the entry prompt.